requirements.txt
flask-restful
python-simplexml